Datasets to PM 2.5
This code is used to visualize the annual PM2.5 emissions data for Mexico, focusing specifically on the importance of these emissions in relation to wildfires. Through a line graph, it presents the evolution of emissions from 2003 to 2024, allowing for the observation of trends and patterns over the years. This topic is particularly relevant given the growing concern over air quality and the impact of wildfires on the environment and public health. The use of tools like pandas for data manipulation and matplotlib for visualization provides an effective way to analyze and communicate information related to this critical environmental issue.